pyproject.toml
[tool.poetry]
name = "optimal-route"
version = "0.1.0"
description = "Technical Task of company WelbeX"
authors = ["NurzhauganovA <2003anvarbek2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
unicorn = "^2.0.1.post1"
uvicorn = "^0.29.0"
django-environ = "^0.11.2"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.29"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
